Created enemy and ground prefabs, added "Player" tag to player
The ground prefab has a NavMeshSurface, walls, a "Lit Room" (Kitchen), and the script attached. The enemy is a simple NavMeshAgent that will freeze when the player enters the lit room (mechanic made for prototype to give the "stalker" feel). Everything can be dragged into the scene along with the player prefab with no further setup in the scene required.