Inventory system and interactable items
created 2 placeholder prefabs for batteries and health, they can be added to the inventory by left clicking them within range (raycast from center of the screen). inventory ui (also a placeholder/template) is opened with the E key, and closed with the x button in the corner. created a central ui manager that handles opening and closing menus, as well as anything else dynamic).